API reference

Blast.compute_T̃ — Function

compute_T̃(ℓ::Number, χ::AbstractArray, R::AbstractArray, kmin::Number, kmax::Number, β::Number; n_cheb::Int = 119, N::Int = 2^(15)+1)

Compute integrals of the Bessels function and the Chebyshev polynomials. This is the precomputation part of the code.

Arguments

ℓ::Number: Multipole order

χ::AbstractArray: Array containing values of the comoving distance.

R::AbstractArray: Array containing values for the R=χ₁/χ₂ variable.

kmin::Number and kmax::Number: Integration range in k.

β::Number: Exponent of the k dependence of the integral. This parameter depends on the combination of tracers: β=2,-2,0 for clustering, cosmic shear and the cross-correlation respectively.

n_cheb::Int: Number of chebyshev polynomials used in the approximation of the power spectra.

N::Int: Number of integration points in k.

Blast.bessel_cheb_eval — Function

bessel_cheb_eval(ℓ::Number, kmin::Number, kmax::Number, χ::AbstractArray, n_cheb::Int, N::Number)

Return the Chebyshev polynomials up to order 'n_cheb+1' and the Bessel function of order 'ℓ' evaluated on the grid of 'N' Chebyshev points in the interval ['kmin', 'kmax'] and on the specified 'χ' points.

Blast.get_clencurt_weights — Function

get_clencurt_weights(kmin::Number, kmax::Number, N::Number)

Return the set of 'N' weights needed to perform the integration with the Clenshaw-Curtis quadrature rule. The weights are rescaled between 'kmin' and 'kmax'.

Blast.get_clencurt_grid — Function

get_clencurt_grid(kmin::Number, kmax::Number, N::Number)

Return the integration points in k. They are a set of 'N' Chebyshev points rescaled between 'kmin' and 'kmax'.

Blast.w_ell_tullio — Function

w_ell_tullio(c::AbstractArray,T::AbstractArray)

Compute the tensor contraction of the chebyshev coefficients of the power spectrum 'c' and the precomputed integrals 'T' to obtain the projected matter densities.

Blast.plan_fft — Function

plan_fft(vals::AbstractArray{<:Number,N})

Create an FFTW real-to-real (R2R) transformation plan for the first axis of a given multidimensional array vals. In practise, the vals array is the power spectrum P(k,χ). The first axis should then contains the wavenumbers k, while the second axis contains the χ information.

Arguments

vals::AbstractArray{<:Number, N}: The input array of any numerical type with N dimensions.

Returns

p::FFTW.rFFTWPlan: A FFTW plan object for transforming vals with the appropriate real to real transformations. This plan can be applied using the * operator (e.g., transformed_vals = p * vals).

Blast.fast_chebcoefs — Function

fast_chebcoefs(vals::AbstractArray{<:Number,N}, plan::FFTW.r2rFFTWPlan)

Efficiently compute the Chebyshev coefficients of a multidimensional array vals using an O(n log n) method. This method leverages FFT-based type-I Discrete Cosine Transform (DCT-I).

Arguments:

vals::AbstractArray{<:Number,N}: A multidimensional array of values for which to compute the Chebyshev coefficients.

plan::FFTW.r2rFFTWPlan: A FFTW plan object for transforming vals with the appropriate real to real transformations. This plan is applied using the * operator (e.g., transformed_vals = p * vals) and performs the DCT of the vals array along the first axis.

Returns:

coefs: An array of the same size as vals, containing the computed Chebyshev coefficients.

Blast.simpson_weight_array — Function

simpson_weight_array(n::Int; T=Float64)

Computes the weights for the Simpson quadrature rule for numerical integration based on the input number of points n.

Arguments

n::Int: The number of points (must be at least 2).

T: (optional) The type of the output array. Defaults to Float64.

Returns

An array of length n with the weights of type T for the Simpson quadrature rule.

Blast.compute_Cℓ — Function

compute_Cℓ(w::AbstractArray{T, 3}, K::AbstractArray{T, 4}, χ::AbstractVector, R::AbstractVector)

Computes the Cℓ's by performing the two outer integrals in χ and R. The integration in χ is performed using the Simpson quadrature rule, while the integration in R is performed using the Clenshaw-Curtis quadrature rule.

Arguments

w::AbstractArray{T, 3}: A 3D array representing the projected matter densities, i.e. the inner integrals in k. The three axis are (ℓ, χ, R).

K::AbstractArray{T, 4}: A 4D array representing the kernel function, with dimensions (i,j,χ,R). i and j are the tomographic bins.

χ::AbstractVector: A 1D array containing the χ values.

R::AbstractVector: A 1D array containing the R values.

Returns

A multi-dimensional array Cℓ with axis (ℓ, i, j) containing the angular power spectrum coefficients in every combination of tomographic bins.