API reference

Blast.FlatΛCDM — Type

FlatΛCDM{T}(; w0 = -1.0, wa = 0.0, H0 = 67.27, Ωm = 0.3156, Ωb = 0.0492, 
    Ωde = 0.6844, As = 2.12107e-9, σ8 = 0.816, Ωk = 0.0, Ωr = 0.0, ns = 0.9645
)

Parameters:

w0: Dark energy equation of state parameter at present time (default: -1).

wa: Time evolution of the dark energy equation of state (default: 0).

H0: Hubble constant in km/s/Mpc (default: 67.27).

Ωm: Matter density parameter (default: 0.3156).

Ωb: Baryon density parameter (default: 0.0492).

Ωde: Dark energy density parameter (default: 0.6844).

As: Scalar amplitude of the primordial power spectrum (default: 2.12107e-9).

σ8: Root-mean-square density fluctuation in spheres of radius 8 Mpc (default: 0.816).

Ωk: Curvature density parameter (default: 0, for flat universe).

Ωr: Radiation density parameter (default: 0, since radiation is negligible at low redshift).

ns: Scalar spectral index (default: 0.9645).

Blast.BackgroundQuantities — Type

BackgroundQuantities{T}(; Hz_array, χz_array)

Parameters:

Hz_array: Array of Hubble parameter values, evaluated on a grid of redshift values (default: zeros(500)).

χz_array: Array of comoving distance values, evaluated on a grid of redshift values (default: zeros(500)).

Blast.AbstractCosmology — Type

AbstractCosmology{T}

An abstract type representing a general cosmological model.

Blast.AbstractBackgroundQuantities — Type

AbstractBackgroundQuantities{T}

An abstract type for background quantities in cosmology, such as the Hubble parameter (H), comoving distance (χ), and the growth factor (D).

Blast.AbstractCosmologicalGrid — Type

AbstractCosmologicalGrid{T}

An abstract type representing a grid on which cosmological quantities are evaluated.

Blast.AbstractCosmologicalProbes — Type

AbstractCosmologicalProbes{T}

An abstract type for the shear, clustering and CMB lensing kernels.

Blast.CosmologicalGrid — Type

CosmologicalGrid{T}(; z_range, k_range)

Parameters:

z_range: Array of redshift values where quantities like the Hubble parameter are evaluated (default: LinRange(0.001, 2.5, 300)).

k_range: Array of wavenumbers for evaluating power spectra or other k-dependent quantities (default: LogSpaced(1e-5, 50, 1000)).

Blast.GalaxyKernel — Type

GalaxyKernel{T} <: AbstractCosmologicalProbes{T}

Represents a galaxy kernel in cosmological calculations, where kernel values are provided for multiple tomographic bins.

Parameters

Kernel::AbstractArray{T, 2}: A 2D array of type T, with dimensions (n_bins, nχ). This stores kernel values for each tomographic bin and a grid of χ values.

Constructors

GalaxyKernel{T}(n_bins::Int, nχ::Int): Creates a GalaxyKernel with the specified n_bins and nχ values, initializing the kernel values to zeros of type T.

GalaxyKernel(n_bins::Int, nχ::Int): Creates a GalaxyKernel with the specified n_bins and nχ values, initializing the kernel values to zeros of type Float64.

Blast.ShearKernel — Type

ShearKernel{T} <: AbstractCosmologicalProbes{T}

Represents a shear kernel used in cosmological lensing calculations. The kernel is defined over multiple tomographic bins.

Parameters

Kernel::AbstractArray{T, 2}: A 2D array of type T, with dimensions (n_bins, nχ). Stores the kernel values for each tomographic bin and a grid of χ values.

Constructors

ShearKernel{T}(n_bins::Int, nχ::Int): Creates a ShearKernel with the specified n_bins and nχ values, initializing the kernel values to zeros of type T.

ShearKernel(n_bins::Int, nχ::Int): Creates a ShearKernel with the specified n_bins and nχ values, initializing the kernel values to zeros of type Float64.

Blast.CMBLensingKernel — Type

CMBLensingKernel{T} <: AbstractCosmologicalProbes{T}

Represents a CMB lensing kernel.

Parameters

Kernel::AbstractArray{T, 2}: A 2D array of type T, with dimensions (n_bins, nχ). Note that CMB Lensing by definition only has a single tomographic bin.

Constructors

CMBLensingKernel{T}(n_bins::Int, nχ::Int): Creates a CMBLensingKernel with the specified n_bins and nχ values, initializing the kernel values to zeros of type T.

CMBLensingKernel(n_bins::Int, nχ::Int): Creates a CMBLensingKernel with the specified n_bins and nχ values, initializing the kernel values to zeros of type Float64.

Blast.compute_T̃ — Function

compute_T̃(ℓ::Number, χ::AbstractArray, R::AbstractArray, kmin::Number, kmax::Number, β::Number; n_cheb::Int = 119, N::Int = 2^(15)+1)

Compute integrals of the Bessels function and the Chebyshev polynomials. This is the precomputation part of the code.

Arguments

ℓ::Number: Multipole order

χ::AbstractArray: Array containing values of the comoving distance.

R::AbstractArray: Array containing values for the R=χ₁/χ₂ variable.

kmin::Number and kmax::Number: Integration range in k.

β::Number: Exponent of the k dependence of the integral. This parameter depends on the combination of tracers: β=2,-2,0 for clustering, cosmic shear and the cross-correlation respectively.

n_cheb::Int: Number of chebyshev polynomials used in the approximation of the power spectra.

N::Int: Number of integration points in k.

Blast.bessel_cheb_eval — Function

bessel_cheb_eval(ℓ::Number, kmin::Number, kmax::Number, χ::AbstractArray, n_cheb::Int, N::Number)

Return the Chebyshev polynomials up to order 'n_cheb+1' and the Bessel function of order 'ℓ' evaluated on the grid of 'N' Chebyshev points in the interval ['kmin', 'kmax'] and on the specified 'χ' points.

Blast.get_clencurt_weights — Function

get_clencurt_weights(kmin::Number, kmax::Number, N::Number)

Return the set of 'N' weights needed to perform the integration with the Clenshaw-Curtis quadrature rule. The weights are rescaled between 'kmin' and 'kmax'.

Blast.get_clencurt_grid — Function

get_clencurt_grid(kmin::Number, kmax::Number, N::Number)

Return the integration points in k. They are a set of 'N' Chebyshev points rescaled between 'kmin' and 'kmax'.

Blast.w_ell_tullio — Function

w_ell_tullio(c::AbstractArray,T::AbstractArray)

Compute the tensor contraction of the chebyshev coefficients of the power spectrum 'c' and the precomputed integrals 'T' to obtain the projected matter densities.

Blast.plan_fft — Function

plan_fft(vals::AbstractArray{<:Number,N})

Create an FFTW real-to-real (R2R) transformation plan for the first axis of a given multidimensional array vals. In practise, the vals array is the power spectrum P(k,χ). The first axis should then contains the wavenumbers k, while the second axis contains the χ information.

Arguments

vals::AbstractArray{<:Number, N}: The input array of any numerical type with N dimensions.

Returns

p::FFTW.rFFTWPlan: A FFTW plan object for transforming vals with the appropriate real to real transformations. This plan can be applied using the * operator (e.g., transformed_vals = p * vals).

Blast.fast_chebcoefs — Function

fast_chebcoefs(vals::AbstractArray{<:Number,N}, plan::FFTW.r2rFFTWPlan)

Efficiently compute the Chebyshev coefficients of a multidimensional array vals using an O(n log n) method. This method leverages FFT-based type-I Discrete Cosine Transform (DCT-I).

Arguments:

vals::AbstractArray{<:Number,N}: A multidimensional array of values for which to compute the Chebyshev coefficients.

plan::FFTW.r2rFFTWPlan: A FFTW plan object for transforming vals with the appropriate real to real transformations. This plan is applied using the * operator (e.g., transformed_vals = p * vals) and performs the DCT of the vals array along the first axis.

Returns:

coefs: An array of the same size as vals, containing the computed Chebyshev coefficients.

Blast.make_grid — Function

make_grid(BackgroundQuantities::BackgroundQuantities, R::Vector{T}) where T

Constructs a grid by multiplying the χz_array from BackgroundQuantities with the vector R.

Arguments

BackgroundQuantities::BackgroundQuantities: An instance of the BackgroundQuantities type that contains the χz_array.

R::Vector{T}: A vector of values to be used in the grid construction, where T can be any type.

Blast.grid_interpolator — Method

grid_interpolator(Probe::Union{GalaxyKernel, ShearKernel}, 
    BackgroundQuantities::BackgroundQuantities, grid::Vector{T}) where T

Interpolates the kernel values for a given grid based on the specified cosmological probes. Returns a 2D array of interpolated kernel values, where rows correspond to the number of bins and columns correspond to the grid points.

Arguments

Probe::Union{GalaxyKernel, ShearKernel}: The kernel data to be interpolated.

BackgroundQuantities::BackgroundQuantities: An instance of the BackgroundQuantities type that contains the χz_array.

grid::Vector{T}: A vector of values where the interpolated kernel values will be evaluated.

Blast.grid_interpolator — Method

grid_interpolator(Probe::CMBLensingKernel, 
    bg::BackgroundQuantities, grid::Vector{T}) where T

Interpolates the kernel values for a given grid based on the specified cosmological probes. Returns a 2D array of interpolated kernel values, where rows correspond to the number of bins and columns correspond to the grid points.

Arguments

Probe::CMBLensingKernel: The kernel data to be interpolated.

bg::BackgroundQuantities: An instance of the BackgroundQuantities type that contains the χz_array.

grid::Vector{T}: A vector of values where the interpolated kernel values will be evaluated.

Blast.get_kernel_array — Method

get_kernel_array(AbstractCosmologicalProbes::GalaxyKernel, 
    BackgroundQuantities::BackgroundQuantities, R::Vector{T}) where T

Obtains the kernel array for the GalaxyKernel probe, with dimensions (bins, nχ, nR).

Arguments

AbstractCosmologicalProbes::GalaxyKernel: An instance of the GalaxyKernel type.

BackgroundQuantities::BackgroundQuantities: An instance of the BackgroundQuantities type.

R::Vector{T}: A vector of values for which the kernel array is to be computed.

Blast.get_kernel_array — Method

get_kernel_array(Probe::ShearKernel, 
    bg::BackgroundQuantities, R::Vector{T}) where T

Obtains the kernel array for the weak lensing probe, with dimensions (bins, nχ, nR) The difference with respect to the galaxy case is that these kernels are divided by $\chi^2$.

Arguments

Probe::ShearKernel: An instance of ShearKernel.

bg::BackgroundQuantities: An instance of the BackgroundQuantities type.

R::Vector{T}: A vector of values for which the kernel array is to be computed.

Blast.get_kernel_array — Method

get_kernel_array(Probe::CMBLensingKernel, 
    bg::BackgroundQuantities, R::Vector{T}) where T

Obtains the kernel array for the CMB lensing probe, with dimensions (1, nχ, nR) Similarly to the weak lensing case, the kernel is divided by $\chi^2$.

Arguments

Probe::CMBLensingKernel: An instance of CMBLensingKernel.

bg::BackgroundQuantities: An instance of the BackgroundQuantities type.

R::Vector{T}: A vector of values for which the kernel array is to be computed.

Blast.combine_kernels — Function

combine_kernels(ProbeA::Union{GalaxyKernel, ShearKernel, CMBLensingKernel}, 
    ProbeB::Union{GalaxyKernel, ShearKernel, CMBLensingKernel}, 
    BackgroundQuantities::BackgroundQuantities, R::Vector{T}) where T

Combines the kernels from two different cosmological probes into a single array. This is needed to perform the integration in the χ-R coordinates.

Arguments

ProbeA::Union{GalaxyKernel, ShearKernel, CMBLensingKernel}: The first cosmological probe to combine.

ProbeB::Union{GalaxyKernel, ShearKernel, CMBLensingKernel}: The second cosmological probe to combine.

BackgroundQuantities::BackgroundQuantities: An instance of the BackgroundQuantities type.

R::Vector{T}: A vector of values used in the combination.

Blast.factorial_frac — Function

factorial_frac(ℓ::Union{Number,Vector{T}}}) where T

Computes the ratio (ℓ+2)!/(ℓ-2)!, needed in the pre-factors of the the angular power spectra.

Arguments

ℓ::Vector{T}: vectors of ℓ values.

Blast.get_ell_prefactor — Method

get_ell_prefactor(ProbeA::GalaxyKernel, ProbeB::GalaxyKernel, ℓ_list::Vector)

Calculates the prefactor for the angular power spectrum when both probes are GalaxyKernel.

Arguments

ℓ_list::Vector: A vector of angular multipole values.

Blast.get_ell_prefactor — Method

get_ell_prefactor(ProbeA::GalaxyKernel, ProbeB::ShearKernel, ℓ_list::Vector)

Calculates the prefactor for the angular power spectrum when probes are GalaxyKernel and ShearKernel.

Arguments

ℓ_list::Vector: A vector of angular multipole values.

Blast.get_ell_prefactor — Method

get_ell_prefactor(ProbeA::ShearKernel, ProbeB::ShearKernel, ℓ_list::Vector)

Calculates the prefactor for the angular power spectrum when both probes are ShearKernel.

Arguments

ℓ_list::Vector: A vector of angular multipole values.

Blast.get_ell_prefactor — Method

get_ell_prefactor(ProbeA::CMBLensingKernel, ProbeB::ShearKernel, ℓ_list::Vector)

Calculates the prefactor for the angular power spectrum when probes are CMBLensingKernel and ShearKernel.

Arguments

ℓ_list::Vector: A vector of angular multipole values.

Blast.get_ell_prefactor — Method

get_ell_prefactor(ProbeA::CMBLensingKernel, ProbeB::CMBLensingKernel, ℓ_list::Vector)

Calculates the prefactor for the angular power spectrum when both probes are CMBLensingKernel.

Arguments

ℓ_list::Vector: A vector of angular multipole values.

Blast.get_ell_prefactor — Method

get_ell_prefactor(ProbeA::CMBLensingKernel, ProbeB::GalaxyKernel, ℓ_list::Vector)

Calculates the prefactor for the angular power spectrum when probes are CMBLensingKernel and GalaxyKernel.

Arguments

ℓ_list::Vector: A vector of angular multipole values.

Blast.simpson_weight_array — Function

simpson_weight_array(n::Int; T=Float64)

Computes the weights for the Simpson quadrature rule for numerical integration based on the input number of points n.

Arguments

n::Int: The number of points (must be at least 2).

T: (optional) The type of the output array. Defaults to Float64.

Returns

An array of length n with the weights of type T for the Simpson quadrature rule.

Blast.get_clencurt_weights_R_integration — Function

get_clencurt_weights_R_integration(N::Int)

Calculate Clenshaw-Curtis quadrature weights for numerical integration over a given interval. The weights are computed for the second half of the Chebyshev points, omitting the zero element. The first weight is halved, although this is an approximation and may benefit from further optimization.

Arguments

N::Int: The number of quadrature points.

Blast.compute_Cℓ — Method

compute_Cℓ(w::AbstractArray{T, 3}, 
           ProbeA::Union{GalaxyKernel, ShearKernel, CMBLensingKernel}, 
           ProbeB::Union{GalaxyKernel, ShearKernel, CMBLensingKernel}, 
           BackgroundQuantities::BackgroundQuantities, 
           R::AbstractVector, 
           ℓ_list::AbstractArray{T,1} = Blast.ℓ) where T

Computes the angular power spectrum Cℓ by performing two outer integrals over χ and R. The Simpson quadrature rule is used for integration over χ, while Clenshaw-Curtis quadrature is used for R.

Arguments

w::AbstractArray{T, 3}: A 3D array representing the projected matter densities, containing the inner integrals over k. The array dimensions are (ℓ, χ, R).

ProbeA::Union{GalaxyKernel, ShearKernel, CMBLensingKernel}: The first cosmological probe kernel type.

ProbeB::Union{GalaxyKernel, ShearKernel, CMBLensingKernel}: The second cosmological probe kernel type.

BackgroundQuantities::BackgroundQuantities: Contains background information, including χz_array for distances in the cosmology.

R::AbstractVector: A 1D array representing the radial grid values for integration.

ℓ_list::AbstractArray{T,1}: An optional 1D array of angular multipole values, defaulting to the global variable Blast.ℓ, the list of ℓ values used for the precomputed part.

Returns

A 3D array Cℓ with dimensions (ℓ, i, j), where i and j represent the tomographic bins. The array contains the computed angular power spectrum coefficients for each combination of ℓ values and tomographic bins.

Blast.compute_Cℓ — Method

compute_Cℓ(w::AbstractArray{T, 3}, 
           K::AbstractArray{T, 4}, 
           bkgq::BackgroundQuantities, 
           weights_χ::AbstractArray{T, 1}, 
           weights_R::AbstractArray{T, 1}, 
           ell_prefactor::AbstractArray{T,1}) where T

Computes the angular power spectrum Cℓ by performing integrals over χ and R for a given set of inputs.

Arguments

w::AbstractArray{T, 3}: A 3D array representing the projected matter densities, containing the inner integrals over k. The array dimensions are (ℓ, χ, R).

K::AbstractArray{T, 4}: A 4D array representing the kernel product K[i, j, χ, R] for the given cosmological probes. It combines the effects of the probes and tomographic bin combinations.

bkgq::BackgroundQuantities: Contains background information, including χz_array for distances in the cosmology.

weights_χ::AbstractArray{T, 1}: A 1D array of weights for Simpson quadrature integration over χ.

weights_R::AbstractArray{T, 1}: A 1D array of weights for Clenshaw-Curtis quadrature integration over R.

ell_prefactor::AbstractArray{T,1}: A 1D array of prefactors dependent on angular multipole values ℓ, combining with other components to form the power spectrum.

Returns

A 3D array Cℓ with dimensions (ℓ, i, j), where i and j represent the tomographic bins. The array contains the computed angular power spectrum coefficients for each combination of ℓ values and tomographic bins.

Blast.compute_adimensional_hubble_factor — Method

compute_adimensional_hubble_factor(z::T, cosmo::FlatΛCDM) -> T

Computes the adimensional Hubble factor E(z) for a given redshift z, using the cosmological parameters from a FlatΛCDM model. The analitycal expression is given by:

\[E(z)=\sqrt{\Omega_m(1+z)^3+\Omega_r(1+z)^4+ \Omega_{de}(1+z)^{3(1+w_0+w_a)}\exp(-3w_a \frac{z}{1+z})+\Omega_k(1+z)^2}\]

Parameters:

z: Redshift at which to evaluate the Hubble factor.

cosmo: A FlatΛCDM cosmological model containing parameters like Ωm, Ωr, Ωde, etc.

Returns:

E_z: The adimensional Hubble factor at redshift z.

Blast.compute_adimensional_hubble_factor — Method

compute_adimensional_hubble_factor(z::T, Ωm::T, Ωr::T, Ωde::T, Ωk::T, w0::T, wa::T) -> T

Computes the adimensional Hubble factor E(z) given the redshift z and individual cosmological parameters. The analitycal expression is given by:

\[E(z)=\sqrt{\Omega_m(1+z)^3+\Omega_r(1+z)^4+ \Omega_{de}(1+z)^{3(1+w_0+w_a)}\exp(-3w_a \frac{z}{1+z})+\Omega_k(1+z)^2}\]

Parameters:

z: Redshift at which to evaluate the Hubble factor.

Ωm: Matter density parameter.

Ωr: Radiation density parameter.

Ωde: Dark energy density parameter.

Ωk: Curvature density parameter.

w0: Dark energy equation of state parameter at the present time.

wa: Time evolution of the dark energy equation of state.

Returns:

E_z: The adimensional Hubble factor at redshift z.

Blast.compute_hubble_factor — Function

compute_hubble_factor(z::T, cosmo::AbstractCosmology) -> T

Computes the Hubble parameter H(z) at a given redshift z using the Hubble constant H0 and the adimensional Hubble factor E(z).

Parameters:

z: Redshift at which to compute the Hubble parameter.

cosmo: A cosmological model that contains H0 and other necessary parameters.

Returns:

H_z: The Hubble parameter at redshift z.

Blast.compute_χ — Function

compute_χ(z::T, cosmo::AbstractCosmology) -> T

Computes the comoving distance χ(z) to a given redshift z by numerically integrating the inverse of the adimensional Hubble factor E(z):

\[\chi(z)=\frac{c}{H_0}\int_0^z \frac{dz'}{E(z')}\]

Parameters:

z: Redshift up to which the comoving distance is computed.

cosmo: A cosmological model containing the necessary parameters (e.g., Ωm, H0).

Returns:

χ_z: The comoving distance at redshift z in units of Mpc.

Blast.evaluate_background_quantities! — Function

evaluate_background_quantities!(grid::CosmologicalGrid, bg::BackgroundQuantities, cosmo::AbstractCosmology)

Populates the BackgroundQuantities struct with values for the Hubble parameter H(z) and comoving distance χ(z) over the redshift range specified by the CosmologicalGrid.

Parameters:

grid: A grid specifying the redshift range over which to evaluate the background quantities.

bg: A mutable struct where the computed H(z) and χ(z) values will be stored.

cosmo: A cosmological model containing the necessary parameters (e.g., H0, Ωm).

Notes:

This function modifies the BackgroundQuantities struct in place by filling the arrays with the computed values.

Blast.compute_kernel! — Method

compute_kernel!(nz::AbstractArray{T, 2}, Probe::GalaxyKernel, z::AbstractArray{T, 1},
                grid::CosmologicalGrid, bg::BackgroundQuantities, 
                cosmo::AbstractCosmology) where T

Computes the galaxy clustering kernel based on a redshift distribution nz and stores it in the GalaxyKernel struct. The kernel is defined as:

\[W_g(\chi) = \frac{H(z)}{c}n(z)\]

Parameters:

nz: A 2D array of type T where each row represents the redshift distribution of galaxies for a specific redshift bin.

z: The redshift grid corresponding to the nz array.

Probe: An instance of GalaxyKernel, in which the computed kernel values for each redshift bin will be stored.

grid: A CosmologicalGrid object specifying the redshift range and grid points for kernel computation.

bg: A struct containing arrays of Hubble parameter (Hz_array) and comoving distance (χz_array), precomputed over the grid.

cosmo: An instance of a cosmological model used to calculate the background quantities if not already provided.

Blast.compute_kernel! — Method

compute_kernel!(nz::AbstractArray{T, 2}, Probe::ShearKernel, z::AbstractArray{T, 1},
                grid::CosmologicalGrid, bg::BackgroundQuantities, 
                cosmo::AbstractCosmology) where T

Computes the weak lensing shear kernel based on a redshift distribution nz and stores it in the ShearKernel struct. The kernel is defined as:

\[W_{\gamma}(\chi) = \frac{3}{2}\frac{H_0^2}{c^2}\Omega_m \frac{\chi}{a(\chi)}\int_{z(\chi)}^{\infty}dz'n(z')\frac{\chi(z')-\chi}{\chi(z')}\]

Parameters:

nz: A 2D array of type T where each row corresponds to the redshift distribution for a specific shear redshift bin.

z: The redshift grid corresponding to the nz array.

Probe: An instance of ShearKernel, where computed kernel values for each redshift bin will be stored.

grid: A CosmologicalGrid object defining the redshift range and grid points for kernel computation.

bg: A struct containing precomputed Hubble parameter (Hz_array) and comoving distance (χz_array) arrays over the grid.

cosmo: An instance of a cosmological model that provides background parameters needed for lensing kernel calculations.

Blast.compute_kernel! — Method

compute_kernel!(Probe::CMBLensingKernel, 
                grid::CosmologicalGrid, bg::BackgroundQuantities, 
                cosmo::AbstractCosmology)

Computes the CMB lensing kernel and stores it in the CMBLensingKernel struct. The kernel is defined as:

\[W_{\kappa}(\chi) = \frac{3}{2}\frac{H_0^2}{c^2}\Omega_m \frac{\chi}{a(\chi)}\frac{\chi^*-\chi}{\chi^*},\]

where $\chi^* = \chi(z_{\mathrm{CMB}} = 1100)$

Parameters:

Probe: An instance of CMBLensingKernel to store the computed kernel values.

grid: A grid specifying the redshift range over which the kernel is computed.

bg: A struct containing precomputed Hubble parameter and comoving distance values.

cosmo: A cosmological model.